Added --carrier-summary-by-phenotype.
This gives carrier/noncarrier counts for all values of given phenotype column
in the phenotype table. So for example:

--carrier-summary-by-phenotype affected will add onto the end of a query
five columns affected_carrier, unaffected carrier, affected_noncarrier,
etc. with an 'unknown' column of all the samples where either the
phenotype or the genotype are unknown.

This will work with any column in the sample table, so for example to
get carrier/noncarrier counts by hair color, just do hair_color instead
of phenotype.

Some collaborators were asking for this kind of summary for prioritizing
variants.